Split out CLI from main code
