+++ bin/ppp
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../lib/cli').run(__filename);
